requirements.txt
torch>=1.12.0
transformers>=4.31.0,!=4.33.2
deepspeed>=0.9.0
datasets>=2.8.0
sentencepiece>=0.1.97
accelerate>=0.15.0
protobuf==3.20.3
torch_geometric
scikit-learn
tqdm
pandas
dgl
numpy
scipy
tensorboard
openai
matplotlib
networkx
pytz
pyyaml
bleurt
bart_score
evaluate